refactor(cloud/scale): Move typed errors into errors.go files
Move typed error declarations from command files into package-
level errors.go files for each cloud/scale subpackage.
This keeps command implementations focused while preserving
existing error behavior and %w wrapping.
