+++ internal/cli/kraft/cloud/scale/reset/errors.go
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2024, Unikraft GmbH and The KraftKit Authors.
+// Licensed under the BSD-3-Clause License (the "License").
+// You may not use this file except in compliance with the License.
+
+package reset
+
+import "errors"
+
+var ErrServiceIdentifierRequired = errors.New("specify a service name or UUID")
